storage: bm_storage_sd: store the number of retries in a single byte
Use a single byte to keep count, to save a few bytes.
Add a range to the Kconfig that controls how many times an operation
can be retried.

+++ subsys/storage/bm_storage/sd/bm_storage_sd.c
@@ -68,8 +68,8 @@ struct {
 		OP_NONE,
 		OP_EXECUTING,
 	} operation_state;
-	/* Number of times an operation has been retried on timeout. */
-	uint32_t retries;
+	/* Number of times an operation has been retried on timeout */
+	uint8_t retries;
 	/* The SoftDevice is enabled. */
 	bool sd_enabled;
 	struct bm_storage_sd_op current_operation;
